+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- added `dc2esmapping` - converts dataclass to elasticsearch mapping
 
 
 0.1.0 (2020-11-13)